+++ packages/riverpod/test/framework/provider_container_test.dart
@@ -8,6 +8,27 @@ import '../utils.dart';
 
 void main() {
   group('ProviderContainer', () {
+    test(
+        'after a child container is disposed, ref.watch keeps working on providers associated with the ancestor container',
+        () async {
+      final container = createContainer();
+      final dep = StateProvider((ref) => 0);
+      final provider = Provider((ref) => ref.watch(dep).state);
+      final listener = Listener<int>();
+      final child = createContainer(parent: container);
+
+      container.listen<int>(provider, listener, fireImmediately: true);
+
+      verifyOnly(listener, listener(0));
+
+      child.dispose();
+
+      container.read(dep).state++;
+      await container.pump();
+
+      verifyOnly(listener, listener(1));
+    });
+
     test('flushes listened providers even if they have no external listeners',
         () async {
       final dep = StateProvider((ref) => 0);